début modif bread grade + creation de post + child edit

+++ src/Controller/GradeController.php
@@ -4,10 +4,12 @@ namespace App\Controller;
 
 use App\Entity\Grade;
 use App\Form\UploadType;
+use App\Form\UploadEditType;
 use App\Repository\GradeRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Doctrine\ORM\EntityManagerInterface;
 
@@ -26,21 +28,84 @@ class GradeController extends AbstractController
     /** 
      * @Route("/grade/browse", name="grade_browse")
      */
-    public function browse (GradeRepository $gradeRepository)
+    public function browse(GradeRepository $gradeRepository)
     {
+        $families = $this->getUser()->getFamilies();
+        $gradesArray = array();
+
+        foreach ($families as $family) {
+            $children = $family->getChildren();
+
+            foreach ($children as $child) {
+                $grades = $child->getGrades()->getValues();
+
+                foreach($grades as $grade) {
+                    array_push($gradesArray, $grade);
+                }
+            }
+        }
         return $this->render('grade/browse.html.twig', [
-            'grades' => $gradeRepository->findAll(),
+            'grades' => $gradesArray,
         ]);
     }
 
     /** 
-     * @Route("/grade/read/{id}", name="grade_read", requirements= {"id": "\d+"})
+     * @Route("/grade/read/{id}", name="grade_read", requirements={"id": "\d+"})
      */
-    public function read (Grade $grade)
+    public function read(Grade $grade)
     {
         return $this->render('grade/read.html.twig', [
             'grade' => $grade,
         ]);
+    }
+
+    /**
+     * @Route("/grade/edit/{id}", name="grade_edit", requirements={"id": "\d+"})
+     */
+    public function edit(Grade $grade, Request $request, EntityManagerInterface $em)
+    {
+        $form = $this->createForm(UploadEditType::class, $grade);
+
+        $form->handleRequest($request);
+        
+        if ($form->isSubmitted() && $form->isValid()){
+            
+            $newFile = $form['file']->getData();
+            
+            if ($newFile != null) {
+                function generateRandomString($length = 10)
+                {
+                    $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+                    $maxLength = strlen($characters);
+                    $randomString = '';
+                    for ($i = 0; $i < $length; $i++) {
+                        $randomString .= $characters[rand(0, $maxLength - 1)];
+                    }
+                    return $randomString;
+                }
+                
+                $fileName = generateRandomString();
+                
+                $directory = 'assets/files/grades/';
+                
+                $finalDirectory = $directory.$fileName.'.jpg';
+                $grade->setFile($finalDirectory);
+                
+                $newFile->move($this->getParameter('grades_directory'), $fileName.'.jpg');
+
+            }
+            
+            $em->persist($grade);
+            $em->flush();
+            return $this->redirectToRoute('grade_browse');
+        }
+            
+            
+        return $this->render('grade/edit.html.twig', [
+            'form' => $form->createView(),
+        ]);
+
+
     }
 
     /**
@@ -59,7 +124,6 @@ class GradeController extends AbstractController
             $name = $form->getData()->getName();
             $file = $form['file']->getData();
             
-
             function generateRandomString($length = 10)
             {
                 $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
@@ -99,12 +163,16 @@ class GradeController extends AbstractController
     /** 
      * @Route("/grade/delete/{id}", name="grade_delete", requirements= {"id": "\d+"})
      */
-    public function delete (Grade $grade, EntityManagerInterface $em)
+    public function delete (Request $request, Grade $grade): Response
     {
+        if ($this->isCsrfTokenValid('delete'.$grade->getId(), $request->request->get('_token'))){
+
+        $em = $this->getDoctrine()->getManager();
         $em->remove($grade);
         $em->flush();
-
-        return $this->redirectToRoute('grade_browse', [
-        ]);
+        
+        $this->addFlash('success', 'bulletin de note supprimé.');
+        }
+        return $this->redirectToRoute('grade_browse');
     }
 }
